⚙️ refactor: rename ai-keymap-handler to ai-handler for consistency
